# docs/openapi.yaml
openapi: 3.0.3
info:
  title: iDME 工业 MiniApp API
  description: |
    华为 iDME 数据建模引擎 - 工业 MiniApp 接口文档。

    ## 快速开始

    ### 推荐创建顺序（依赖链）

    **Level 0（无依赖）**：Unit（计量单位）, BusinessPartner（往来单位）, EquipmentClassfication（设备分类）, PartClassfication（物料分类）

    **Level 1**：Department（部门）, Employee（员工）

    **Level 2**：EquipmentModel（设备型号）, Location（位置）

    **Level 3**：Equipment（设备）, Part（物料）

    **Level 4**：WorkingProcedure（工序）, WorkingPlan（工艺路线）

    **Level 5**：BOMItem（BOM项）, PlanProcedureLink（计划-工序关联）, ProcedureEquipmentLink（工序-设备关联）, ProcedurePartLink（工序-物料关联）, EquipmentSparePartLink（设备-备件关联）

    ### 注意事项
    - **引用字段**（如 `manufacturerName`, `category`）需要传入对应模块的 **ID**，不是文本名称
    - **Part** 和 **WorkingPlan** 是版本对象，`WorkingPlan.productPart` 引用 Part 时使用 `master.id` 而非实例 `id`
    - **枚举值**使用英文（如 `Standby`, `Active`, `Machinery`）
    - 所有接口的 example 字段已填充真实可用的示例ID，可直接复制使用

    ### 示例数据
    - 固定ID文档: `/docs/openapi-examples-ids.md`
    - 测试文件: `/src/test/java/com/tsukilc/idme/integration/OpenApiExamplesDataTest.java`

    ---

    - 管理界面: http://99.suyiiyii.top:8003/rdm_a1b52ff379ee46ed8928d7f2ceb908f6_app/services/index.html
    - 健康检查: http://99.suyiiyii.top:8003/rdm_a1b52ff379ee46ed8928d7f2ceb908f6_app/services/v1/health

    认证: Basic Auth 用户名 sysadmin / 密码见项目说明。
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://99.suyiiyii.top:8003/rdm_a1b52ff379ee46ed8928d7f2ceb908f6_app/services
    description: iDME SDK 服务
security:
  - basicAuth: []
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: 用户名 sysadmin，密码见项目文档
  parameters:
    pageNum:
      name: pageNum
      in: query
      description: 页码，从 1 开始
      schema:
        type: integer
        default: 1
    pageSize:
      name: pageSize
      in: query
      description: 每页大小
      schema:
        type: integer
        default: 20
    idPath:
      name: id
      in: path
      required: true
      description: 主键 ID
      schema:
        type: string
  schemas:
    ApiResponse:
      type: object
      properties:
        result:
          type: string
          example: SUCCESS
        data: {}
        errors:
          type: array
          items:
            type: string
      description: 统一响应包装
...
    PartnerRef:
      type: object
      description: 往来单位引用
      properties:
        id: { type: string }
        partnerName: { type: string }
    PartRef:
      type: object
      description: 物料引用（含名称）
      properties:
        id: { type: string }
        partName: { type: string }
...
    ProcedureRef:
      type: object
      description: 工序引用
      properties:
        id: { type: string }
        procedureName: { type: string }
    ProcedureRefWithCode:
      type: object
      description: 工序引用（含编码）
      properties:
        id: { type: string }
        procedureCode: { type: string }
        procedureName: { type: string }
...
    PlanProcedureItem:
      type: object
      description: 工艺路线工序项（包含关联信息和工序详情）
      properties:
        linkId: { type: string, description: 关联ID }
        sequence: { type: integer, description: 工序顺序 }
        standardDurationMin: { type: number, description: 标准工时(分钟) }
        procedureId: { type: string, description: 工序ID }
        procedureCode: { type: string, description: 工序编号 }
        procedureName: { type: string, description: 工序名称 }
        steps: { type: string, description: 操作步骤 }
        status: { type: string, description: 状态 }
        mainProductionEquipment: { type: string, description: 主生产设备ID }
        mainInspectionEquipment: { type: string, description: 主检测设备ID }
        operatorRef: { type: string, description: 操作人员ID }
        remarks: { type: string, description: 备注 }
...
    EquipmentDetail:
      type: object
      description: 设备详情（含关联对象）
      properties:
        id: { type: string }
        equipmentCode: { type: string }
        equipmentName: { type: string }
        manufacturerName:
          oneOf:
            - { type: string }
            - { $ref: '#/components/schemas/PartnerRef' }
        brand: { type: string }
        modelSpec: { type: string }
        status: { type: string, enum: [运行, 待机, 维修, 停机, 报废] }
        techParams: { type: object, additionalProperties: true }
        createTime: { type: string, format: date-time }
        lastModifiedTime: { type: string, format: date-time }
    EquipmentListItem:
      type: object
      description: 设备列表项
      properties:
        id: { type: string }
        equipmentCode: { type: string }
        equipmentName: { type: string }
        status: { type: string }
        locationText: { type: string }
...
    WorkingPlanDetail:
      type: object
      description: 工艺路线详情（含工序、产品物料）
      properties:
        id: { type: string }
        planCode: { type: string }
        planName: { type: string }
        businessVersion: { type: string }
        productPart: { $ref: '#/components/schemas/PartRef' }
        description: { type: string }
        procedures:
          type: array
          items: { $ref: '#/components/schemas/PlanProcedureItem' }
        status: { type: string, enum: [Draft, Released, Obsolete] }
...
    EquipmentCreate:
      type: object
      required: [equipmentCode, equipmentName, manufacturerName]
      properties:
        equipmentCode:
          type: string
          description: 设备编码（唯一）
          example: "EQ-PROD-DEMO"
        equipmentName:
          type: string
          description: 设备名称
          example: "CNC加工中心A"
        manufacturerName:
          type: string
          description: 生产厂家ID（-> BusinessPartner模块，partnerType=Machinery）
          example: "2000000001"
        brand:
          type: string
          description: 品牌
          example: "华为"
        modelSpec:
          type: string
          description: 型号规格
        equipmentModelRef:
          type: string
          description: 设备型号ID（-> EquipmentModel模块，SDK必填）
          example: "5000000001"
        supplierName:
          type: string
          description: 供应商ID（-> BusinessPartner模块，SDK必填）
          example: "2000000002"
        productionDate:
          type: string
          format: date
          description: 生产日期
        serviceLifeYears:
          type: integer
          description: 使用年限
        depreciationMethod:
          type: string
          enum: [NoDepreciation, StraightLine, Double, SumOfYear]
          description: "折旧方式：不折旧/直线法/双倍余额递减法/年数总和法"
          example: "StraightLine"
        locationRef:
          type: string
          description: 位置ID（-> Location模块，SDK必填）
          example: "5000000002"
        locationText:
          type: string
          description: 位置文本描述
        status:
          type: string
          enum: [Standby, InOperation, UnderMaintenance, Shutdown, Scrap]
          description: "设备状态：待机/运行中/维修中/停机/报废"
          example: "Standby"
        serialNumber:
          type: string
          description: 序列号
          example: "SN-PROD-001"
        category:
          type: string
          description: 设备分类ID（-> EquipmentClassfication模块）
          example: "3000000001"
        techParams:
          type: object
          description: 技术参数（JSON扩展）
          example: {"精度": "0.001mm", "功率": "7.5kW"}
        remarks:
          type: string
          description: 备注
          example: "示例设备"
...
    WorkingPlan:
      type: object
      properties:
        id: { type: string }
        masterId: { type: string }
        planCode: { type: string }
        planName: { type: string }
        businessVersion: { type: string }
        versionNumber: { type: string }
        productPart: { type: object }
        description: { type: string }
        operatorUser: { type: string }
        operatorRef: { type: string }
        operateTime: { type: string, format: date-time }
        equipmentUsage: { type: object }
        status: { type: string, enum: [Draft, Released, Obsolete] }
        remarks: { type: string }
        procedures: { type: array, items: {} }
    WorkingPlanCreate:
      type: object
      required: [planCode, planName]
      properties:
        planCode:
          type: string
          description: 工艺路线编码
          example: "PLAN-DEMO"
        planName:
          type: string
          description: 工艺路线名称
          example: "中心轮零件加工"
        businessVersion:
          type: string
          description: 业务版本号（版本对象必填）
          example: "1.0"
        productPart:
          type: string
          description: "产品物料ID（-> Part.master.id，⚠️注意：引用Part的master ID，不是实例ID！）"
          example: "864256962114625536"
        description:
          type: string
          description: 描述
          example: "中心轮零件加工工艺路线"
        operatorUser:
          type: string
          description: 操作员用户ID（-> User模块，可选）
        operatorRef:
          type: string
          description: 操作员ID（-> Employee模块，SDK必填）
          example: "4000000002"
        operateTime:
          type: string
          format: date-time
          description: 操作时间
        equipmentUsage:
          type: object
          description: 设备使用情况（JSON扩展）
        status:
          type: string
          description: 工艺路线状态
        remarks:
          type: string
          description: 备注
...
    PlanProcedureLinkCreate:
      type: object
      required: [plan, procedure, sequenceNo]
      properties:
        plan:
          type: string
          description: 工艺路线ID（-> WorkingPlan模块）
          example: "8000000002"
        procedure:
          type: string
          description: 工序ID（-> WorkingProcedure模块）
          example: "8000000001"
        sequenceNo:
          type: integer
          description: 序号
          example: 1
        standardDurationMin:
          type: number
          description: 标准工时（分钟）
          example: 60
        requirement:
          type: string
          description: 工艺要求
    PlanProcedureLinkBatch:
      type: object
      required: [planId, procedures]
      properties:
        planId: { type: string }
        procedures:
          type: array
          items:
            type: object
            properties:
              procedureId: { type: string }
              sequenceNo: { type: integer }
              standardDurationMin: { type: number }
              requirement: { type: string }
    PlanProcedureLink:
      type: object
      description: 计划-工序关联响应
      properties:
        id: { type: string }
        plan: { type: string, description: 计划ID }
        procedure: { type: string, description: 工序ID }
        name: { type: string, description: 名称 }
        description: { type: string, description: 描述 }
        standardDurationMin: { type: number, description: 标准时长（分钟） }
        sequenceNo: { type: integer, description: 序号 }
        requirement: { type: string, description: 要求 }
    ProcedureEquipmentLinkCreate:
      type: object
      required: [procedure, equipment1]
      properties:
        procedure:
          type: string
          description: 工序ID（-> WorkingProcedure模块）
          example: "8000000001"
        equipment1:
          type: string
          description: 设备ID（-> Equipment模块）
          example: "6000000001"
        role:
          type: string
          enum: [Production, Inspection, Auxiliary]
          description: "设备角色：生产/检测/辅助"
          example: "Production"
        plannedStart:
          type: string
          format: date-time
          description: 计划开始时间
        plannedEnd:
          type: string
          format: date-time
          description: 计划结束时间
        actualStart:
          type: string
          format: date-time
          description: 实际开始时间
        actualEnd:
          type: string
          format: date-time
          description: 实际结束时间
        remarks:
          type: string
          description: 备注
    ProcedureEquipmentLink:
      type: object
      description: 工序-设备关联响应
      properties:
        id: { type: string }
        procedure: { type: string, description: 工序ID }
        procedureName: { type: string, description: 工序名称 }
        equipment1: { type: string, description: 设备ID }
        equipmentName: { type: string, description: 设备名称 }
        role:
          type: string
          enum: [Production, Inspection, Auxiliary]
          description: 设备角色
        plannedStart: { type: string, format: date-time }
        plannedEnd: { type: string, format: date-time }
        actualStart: { type: string, format: date-time }
        actualEnd: { type: string, format: date-time }
        remarks: { type: string }
        createTime: { type: string, format: date-time }
        lastUpdateTime: { type: string, format: date-time }
...
    ProcedureStatusUpdate:
      type: object
      properties:
        status: { type: string }
        actualStart: { type: string, format: date-time }
        actualEnd: { type: string, format: date-time }
    SequenceUpdate:
      type: object
      properties:
        sequenceNo: { type: integer }
...
paths:
...
  /api/equipment/{id}:
    parameters: [{ $ref: '#/components/parameters/idPath' }]
    get:
      tags: [设备管理]
      summary: 查询设备详情
      operationId: getEquipment
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EquipmentDetail'
    put:
      tags: [设备管理]
      summary: 更新设备
      operationId: updateEquipment
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EquipmentCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    delete:
      tags: [设备管理]
      summary: 删除设备
      operationId: deleteEquipment
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...
  /api/working-plan:
    get:
      tags: [工艺路线]
      summary: 分页查询工艺路线列表
      operationId: listWorkingPlan
      parameters:
        - $ref: '#/components/parameters/pageNum'
        - $ref: '#/components/parameters/pageSize'
        - name: planName
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          total: { type: integer }
                          pageNum: { type: integer }
                          pageSize: { type: integer }
                          records:
                            type: array
                            items: { $ref: '#/components/schemas/WorkingPlan' }
    post:
      tags: [工艺路线]
      summary: 创建工艺路线
      operationId: createWorkingPlan
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkingPlanCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WorkingPlan'
  /api/working-plan/{id}:
    parameters: [{ $ref: '#/components/parameters/idPath' }]
    get:
      tags: [工艺路线]
      summary: 查询工艺路线详情
      operationId: getWorkingPlan
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WorkingPlanDetail'
    put:
      tags: [工艺路线]
      summary: 更新工艺路线
      operationId: updateWorkingPlan
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkingPlanCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    delete:
      tags: [工艺路线]
      summary: 删除工艺路线
      operationId: deleteWorkingPlan
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...
  /api/working-plan/{id}/procedures:
    parameters: [{ $ref: '#/components/parameters/idPath' }]
    get:
      tags: [工艺路线]
      summary: 查询工艺路线的所有工序
      operationId: getWorkingPlanProcedures
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/PlanProcedureItem' }
...
  /api/plan-procedure-link:
    post:
      tags: [工艺工序关联]
      summary: 为工艺路线添加工序
      operationId: createPlanProcedureLink
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlanProcedureLinkCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PlanProcedureLink'
  /api/plan-procedure-link/batch:
    post:
      tags: [工艺工序关联]
      summary: 批量添加工序（按顺序）
      operationId: batchCreatePlanProcedureLink
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlanProcedureLinkBatch' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/plan-procedure-link/by-plan/{planId}:
    get:
      tags: [工艺工序关联]
      summary: 查询工艺路线的工序列表（已排序）
      operationId: getPlanProcedureLinksByPlan
      parameters:
        - name: planId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/PlanProcedureItem' }
  /api/plan-procedure-link/{id}:
    parameters: [{ $ref: '#/components/parameters/idPath' }]
    put:
      tags: [工艺工序关联]
      summary: 更新工序顺序
      operationId: updatePlanProcedureLinkSequence
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SequenceUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    delete:
      tags: [工艺工序关联]
      summary: 删除工艺工序关联
      operationId: deletePlanProcedureLink
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/procedure-equipment-link:
    post:
      tags: [工序设备关联]
      summary: 为工序关联设备
      operationId: createProcedureEquipmentLink
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProcedureEquipmentLinkCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProcedureEquipmentLink'
  /api/procedure-equipment-link/{id}/actual-time:
    parameters: [{ $ref: '#/components/parameters/idPath' }]
    patch:
      tags: [工序设备关联]
      summary: 更新设备实际使用时间
      operationId: updateProcedureEquipmentActualTime
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProcedureStatusUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/procedure-equipment-link/by-procedure/{procedureId}:
    get:
      tags: [工序设备关联]
      summary: 查询工序使用的所有设备
      operationId: getProcedureEquipmentByProcedure
      parameters:
        - name: procedureId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            id: { type: string }
                            procedure: { $ref: '#/components/schemas/ProcedureRef' }
                            equipment1: { $ref: '#/components/schemas/EquipmentListItem' }
                            role: { type: string }
                            plannedStart: { type: string, format: date-time }
                            plannedEnd: { type: string, format: date-time }
                            actualStart: { type: string, format: date-time }
                            actualEnd: { type: string, format: date-time }
  /api/procedure-equipment-link/by-equipment/{equipmentId}:
    get:
      tags: [工序设备关联]
      summary: 查询设备参与的所有工序
      operationId: getProcedureEquipmentByEquipment
      parameters:
        - name: equipmentId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            id: { type: string }
                            procedure: { $ref: '#/components/schemas/ProcedureRefWithCode' }
                            equipment1: { $ref: '#/components/schemas/EquipmentListItem' }
                            role: { type: string }
  /api/procedure-equipment-link/{id}:
    parameters: [{ $ref: '#/components/parameters/idPath' }]
    delete:
      tags: [工序设备关联]
      summary: 删除工序设备关联
      operationId: deleteProcedureEquipmentLink
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...